requirements.txt
Flask==3.0.3
python-telegram-bot==21.4